Eight Bit Barrel Shifter
Barrel shifters are quite useful in microprocessors since they simplify some of the ALU operations which are costly otherwise. Shifting and Rotating are two basic types of operations carried out by these devices. Shifter can involve the Carry bit of the CPU as well.

This design has being inspired by the features available in the barrel shifter available in the ARM Cortex M3 architecture and the rotation operations available in the Microchip PIC16F architecture.

To keep the design simple, this entity is designed as an eight bit barrel shifter. However, the design is expected to be extensible to higher widths as well.